Reduce stack slot count to i32
i64 is an outrageous number of slots for a single stack frame to have.
Using i32 exceeds the virtual address space on 32bit and would
accommodate 32GiB of roots on 64bit. Reduce to i32 which saves space and
fixes alignment on 32bit.

+++ runtime/alloc/ShadowStackEntry.h
@@ -16,7 +16,7 @@ namespace alloc
 struct ShadowStackEntry
 {
 	ShadowStackEntry *next;
-	std::uint64_t cellCount;
+	std::uint32_t cellCount;
 	AnyCell* roots[];
 };
